// src/app/shared/interfaces/post-condutor.model.ts
export interface PostCondutorMV {
  status?: string;
  nome?: string;
  dataNascimento?: number;
  rg?: string;
  orgaoEmissor?: string;
  cpf?: string;
  nomeMae?: string;
  nomePai?: string;
  telefone?: string;
  celular?: string;
  email?: string;
  cep?: string;
  logradouro?: string;
  numero?: string;
  complemento?: string;
  bairro?: string;
  uf?: string;
  municipioId?: number;
  numeroRegistro?: string;
  habilitacaoCategoriaId?: number;
  dataValidade?: number;
  dataPrimeiraHabilitacao?: number;
  municipioEmissorId?: number;
  dataEmissao?: number;
  observacoes?: string;
  clienteId?: number;
  perfil?: any;
  assinaMulta?: boolean;
  condutorFormaContatoId?: number;
  usuarioId?: number;
  informaEndereco?: boolean;
  informaHabilitacao?: boolean;
  criarUsuarioPortalCliente?: boolean;
  grupoEconomico: any;
  clientes: any[];
  regionais: any[];
  centrosCusto: any[];
  produtos: any[];
  atividades: any[];
  aprovacaoNivel1: boolean;
  aprovacaoNivel2: boolean;
  enderecoCrlv: any;
}